move backend class to testFixtures
